feat(dice/m11b-3): SuccessRule -> enum {Numeric (default), HasSymbol}
impl Default is hand-written rather than derived: derive(Default)'s
#[default] attribute only supports a unit enum variant, and Numeric
carries fields (brief specified a derive that doesn't compile).

+++ src/server/src/dice/eval/crit.rs
@@ -53,7 +53,7 @@ mod tests {
 
     fn cfg(cs: Option<CritSuccess>, cf: Option<CritFail>) -> SuccessConfig {
         SuccessConfig {
-            success: SuccessRule {
+            success: SuccessRule::Numeric {
                 comp: Comparator::Gte,
                 target: 7,
             },
